backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
motor>=3.6.0
python-dotenv>=1.0.1
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<5.0
email-validator>=2.2.0
google-generativeai>=0.7.0
